[PetMatch-16741]: Не работают параметры пагинации при получении карточек животных в админке
[PetMatch-16741]: Не работают параметры пагинации при получении карточек животных в админке

+++ src/main/java/org/nsu/admin/dto/AdminUserListResponse.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class AdminUserListResponse {
 
     @Schema(description = "Total number of records")
-    private Long total;
+    private Integer total;
 
     @Schema(description = "List of users")
     private List<AdminUserDto> users;
